Moved Autos tested to work
