+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/autonomous/routines/RightOfDriverLeftParkRightSpecimenAuton.java
@@ -1,4 +1,4 @@
-package org.firstinspires.ftc.teamcode.autonomous;
+package org.firstinspires.ftc.teamcode.autonomous.routines;
 
 
 import com.acmerobotics.dashboard.config.Config;
@@ -11,19 +11,20 @@ import com.acmerobotics.roadrunner.ftc.Actions;
 import com.qualcomm.robotcore.eventloop.opmode.Autonomous;
 import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
 
+import org.firstinspires.ftc.teamcode.autonomous.MecanumDrive;
 import org.firstinspires.ftc.teamcode.autonomous.subsystems.Arm;
 import org.firstinspires.ftc.teamcode.autonomous.subsystems.Intake;
 import org.firstinspires.ftc.teamcode.autonomous.subsystems.Wrist;
 
 @Config
-@Autonomous(name="Driver: Left, Park: Left, Specimen: Right", group="Autonomous")
+@Autonomous(name="Driver: Right, Park: Left, Specimen: Right", group="Autonomous")
 
 
-public class LeftOfDriverParkLeftSpecimenRightAuton extends LinearOpMode {
+public class RightOfDriverLeftParkRightSpecimenAuton extends LinearOpMode {
 
     @Override
     public void runOpMode() {
-        Pose2d startPose = new Pose2d(-12, -63.5, Math.toRadians(90));
+        Pose2d startPose = new Pose2d(12, -63.5, Math.toRadians(90));
 
         MecanumDrive drive = new MecanumDrive(hardwareMap, startPose);
 
@@ -31,15 +32,22 @@ public class LeftOfDriverParkLeftSpecimenRightAuton extends LinearOpMode {
         Wrist wrist = new Wrist(hardwareMap);
         Intake intake = new Intake(hardwareMap);
 
+//        Action approachSubmersible = drive.actionBuilder(startPose)
+//                .strafeTo(new Vector2d(9, -39))
+//                .build();
+//                //move arm to specimen hanging position
+
         Action moveToSubmersibleToScoreSubmersible = drive.actionBuilder(startPose)
                 .strafeTo(new Vector2d(5, -42))
                 .build();
+                //move arm down a little bit
 
-        Action reverseAndScoreInSubmersibleSlight = drive.actionBuilder(new Pose2d(5, -42, 90))
+
+        Action reverseAndScoreInSubmersibleSlight = drive.actionBuilder(new Pose2d(5, -42, Math.toRadians(90)))
                 .strafeTo(new Vector2d(5, -46))
                 .build();
 
-        Action reverseAndScoreInSubmersibleFull = drive.actionBuilder(new Pose2d(5, -46, 90))
+        Action reverseAndScoreInSubmersibleFull = drive.actionBuilder(new Pose2d(5, -46, Math.toRadians(90)))
                 .strafeTo(new Vector2d(5, -50))
                 .build();
